+++ CHANGELOG.md
@@ -1,6 +1,5 @@
 ### ✨ Features
-- [Config] 新增了 `show_discord_avatar` 用于控制转发消息时是否携带用户的 Discord 头像（默认开启）
-- [Config] 新增了 `sync_edit_delete` 用于控制是否同步 Discord 侧编辑 / 删除消息事件到 QQ 侧（默认关闭）
+- [Config] 新增了 `show_discord_channel_name` 用于控制转发消息到 QQ 时是否显示 Discord 的频道名称（默认关闭）
 
 ### 🐞 Bug Fixes
 - None